Fix: Section 컴포턴트 별로 분리
